Refactor the backend into multiple files
This change makes it possible to use multiple source files
on the server side. `server.js` has been refactored into
multiple files, which should make it easier to find things
moving forward.
