+++ client/src/components/ConfirmDeleteBox.js
